MEGAMOVE
Moving to the rottexpr codebase.

I'm dumping all the previous code and moving onto the rottexpr code as a base because upon closer examination, it's literally doing everything I wanted to initially do in the first place with drott.

I'll reimplement the features I added previously from the rott94-based code and then continue working on this new base.